#75635 - add unit tests for route and replace
